fix namespaces and undo class renames
