+++ ExternalResources/FPDFGroupPages.php
@@ -1,7 +1,7 @@
 <?php
-namespace ExternalResources\FPDF;
+namespace FPDF\ExternalResources;
 
-class ExternalResourcesFPDFGroupPages extends FPDF
+class FPDFGroupPages extends FPDF
 {
     private $NewPageGroup;   // variable indicating whether a new group was requested
     private $PageGroups;     // variable containing the number of pages of the groups
