+++ FPDFGroupPages.php
@@ -1,8 +1,8 @@
 <?php
-namespace fpdf\FPDFGroupPages;
-use fpdf\vendorFPDF;
+namespace FPDF;
+use FPDF\FPDF;
 
-class vendorFPDFGroupPages extends vendorFPDF
+class PDFGroupPages extends FPDF
 {
     private $NewPageGroup;   // variable indicating whether a new group was requested
     private $PageGroups;     // variable containing the number of pages of the groups
