+++ fpdf.php
@@ -1,5 +1,5 @@
 <?php
-namespace fpdf;
+namespace FPDF;
 
 /*******************************************************************************
  * Software: FPDF                                                               *
